Add content-editor component
Add also the `previewHandler` function to be called on the click handler
for the content-editor component. It will fetch the pluginMap content
where we will store the Parser plugin instances markdown content, on the
pluginMarkdownMap array.

As this array will have the same index for the matches of the plugin
RegEx it will be easy to build the final markdown later.