Create basic functionality for the contend-editor
Add the mapPluginMarkdown call to work with the previewHandler as a way
to extract the markdown content from the Parser plugin instances. Where
they will receive a new instance of the mapPluginMarkdown function bound
to the pluginMap object. It will then receive the markdown content so it
can later be used.