Update plugins so they call getMarkdown on each render pass
The getMarkdown call will now be executed on each render pass. This to
ensure that the plugin calls it early on for setting a default markdown
content, and also simplifies the process of updating the markdown
content. Since the plugin design patter is to have a state variable to
manage between edit mode and not edit mode.