Update ImagePlugin to become a React.Component
Now the ImagePlugin follows a similar, but improved, design to the
TextPlugin. It has already some functions for the intended drag and drop
and file upload functions, to be fully implemented on a future branch.

It will take an image url and use it as the src. The improvements are
where the component manages it's style with the isPreviewing prop to
eliminate the borders and show a cleaner slate.