Make the markdown property for the plugin more flexible
Instead of receiving a string to represent the markdown to be used, the
Parser now accepts it to be a funtion. So the plugin can implement a
function to read from the plugin data, and generate the desired
markdown.

This is specially useful for plugins that will have assets to be
uploaded and later have their markdown updated accordingly