Create the store
The store is a factory that will return an object to be used to persist
data. For this initial implementation we have the save method only.

The store factory will take an adapter. Which will consist of an object
that will provide the store with the desired persistence implementation.
It consists of a normalizer function, serializer funtion and a config
object.

By default the store comes with a dropbox and HTTP adapter. The dropbox
adapter is still on experimental stages. The HTTP adapter works with the
server's asset API endpoint, and is the default adapter to be used.