Update the pluginDataMap object logic
Instead of re-instanciating it over every new render pass, we now cache
it's previous state. This should allow us to avoid expensive string
parsing and array mapping calls. Whenever an existing plugin instance is
found on the array, it is returned straight away and used to parse the
plugin/react component.