Use require instead of import for non-critical dependencies
Use require instead of declaring the md-plugin MarkdownEditor and Marked
dependencies with ES6 import. The reason for that is that we spent
around 60 milliseconds on the parsing and initialization of the
md-plugin dependencies alone on the first render, and it would
potentially not even be used, or used much later down on the line.

With that we were able to reduce the first render time in aproximately
70seconds (initial download size + parsing).