Update PluginConstructor to indicate plugin data updates
Now, whenever the pluginConstructor receives a isDataUpdated props, and
it has a 'true' value, it will see if the pluginData.isPluginUpdated
also matches with that value. If it does, the plugin will show a small
animation, giving the user feedback on the data saved.

And add an animation for when the plugin accepted the file to be saved,
on the onDrop call.

Also, rework a bit the main.scss for a happier linter.